requirements.txt
langchain-community
langchain
langchain-core
langchain-groq
sqlalchemy
mysqlclient
pandas
langchain-huggingface
faiss-cpu
streamlit 
mem0ai
flask
twilio
fastapi
uvicorn
ngrok
python-multipart
python-telegram-bot
litellm
opik
SpeechRecognition